CLEANUP of imports and unnecessary classes.
MAJOR CHANGE IN GAME AND EngineAgent Classes.

Now agents are initialised by the Engine Agent as a means to clarify code.
Therefore VisAgent class does not require an AgentController in the constructor anymore.

The way engine now performs actions has been cleaned up to follow better JADE's examples of behaviours. This means that unnecessary loops have been replaced with behaviours and where we cannot control the lifecycle of these behaviours we have setup behaviours which are timed and can be easily stopped.